+++ crates/fluvio-dataplane-protocol/src/error_code.rs
@@ -14,6 +14,7 @@ use crate::smartstream::SmartStreamRuntimeError;
 
 #[repr(i16)]
 #[derive(thiserror::Error, Encoder, Decoder, PartialEq, Debug, Clone)]
+#[non_exhaustive]
 pub enum ErrorCode {
     #[fluvio(tag = -1)]
     #[error("An unknown server error occurred")]
@@ -102,9 +103,14 @@ pub enum ErrorCode {
     #[fluvio(tag = 5000)]
     #[error("an error occurred while managing a connector")]
     ManagedConnectorError,
+
     #[fluvio(tag = 5001)]
     #[error("the managed connector was not found")]
     ManagedConnectorNotFound,
+
+    #[fluvio(tag = 5002)]
+    #[error("an error occurred while managing a connector")]
+    ManagedConnectorAlreadyExists,
 
     // Smart Module Errors
     #[fluvio(tag = 6000)]