+++ crates/fluvio-spu-schema/src/server/stream_fetch.rs
@@ -112,11 +112,13 @@ impl Default for SmartModuleInvocationWasm {
 pub enum SmartStreamKind {
     Filter,
     Map,
-    #[fluvio(min_version = ARRAY_MAP_WASM_API)]
-    ArrayMap,
     Aggregate {
         accumulator: Vec<u8>,
     },
+    #[fluvio(min_version = ARRAY_MAP_WASM_API)]
+    ArrayMap,
+    #[fluvio(min_version = ARRAY_MAP_WASM_API)]
+    FilterMap,
 }
 
 impl Default for SmartStreamKind {